pyproject.toml
[project]
name = "chain-ollama-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.4.1",
    "duckduckgo-search>=7.5.2",
    "openai-agents>=0.0.4",
]

[tool.mypy]
mypy_path = "src"